src/agentica-mini/pyproject.toml
[project]
name = "agentica"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "openai",
    "ipython",
    "typeguard>=4.4.4",
    "python-dotenv>=1.1.1",
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.403",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentica"]
